lots of work on compression

+++ src/core/criteria.rs
@@ -6,7 +6,7 @@ use crate::prelude::*;
 pub type PartitionCriterion<T, U> = Box<dyn (Fn(&Cluster<T, U>) -> bool) + Send + Sync>;
 
 /// A `Cluster` must have a `depth` lower than the given threshold for it to be partitioned.
-pub fn max_depth<T: Number, U: Number>(threshold: u8) -> PartitionCriterion<T, U> {
+pub fn max_depth<T: Number, U: Number>(threshold: usize) -> PartitionCriterion<T, U> {
     Box::new(move |cluster: &Cluster<T, U>| cluster.depth() < threshold)
 }
